Update useKubes to remove uneeded destructured elements
