+++ packages/teleterm/src/ui/DocumentCluster/ClusterResources/Kubes/useKubes.ts
@@ -14,41 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { AttemptStatus } from 'shared/hooks/useAsync';
-
 import { Kube, ServerSideParams } from 'teleterm/services/tshd/types';
 import { useAppContext } from 'teleterm/ui/appContextProvider';
 import { useClusterContext } from 'teleterm/ui/DocumentCluster/clusterContext';
+import { makeKube } from 'teleterm/ui/services/clusters';
 
 import { useServerSideResources } from '../useServerSideResources';
-
-function getEmptyTableText(status: AttemptStatus) {
-  switch (status) {
-    case 'error':
-      return 'Failed to fetch kubernetes clusters.';
-    case '':
-      return 'Searching…';
-    case 'processing':
-      return 'Searching…';
-    case 'success':
-      return 'No kubernetes clusters found.';
-  }
-}
 
 export function useKubes() {
   const appContext = useAppContext();
   const ctx = useClusterContext();
-  const {
-    fetchAttempt,
-    updateQuery,
-    agentFilter,
-    prevPage,
-    nextPage,
-    updateSearch,
-    onAgentLabelClick,
-    updateSort,
-    pageCount,
-  } = useServerSideResources<Kube>(
+  const { fetchAttempt, ...serversideResources } = useServerSideResources<Kube>(
     { fieldName: 'name', dir: 'ASC' }, // default sort
     (params: ServerSideParams) => appContext.resourcesService.fetchKubes(params)
   );
@@ -56,21 +32,8 @@ export function useKubes() {
   return {
     connect: ctx.connectKube,
     fetchAttempt,
-    kubes: fetchAttempt.data?.agentsList,
-    agentFilter,
-    updateQuery,
-    updateSearch,
-    onAgentLabelClick,
-    pageCount,
-    disabledRows: fetchAttempt.status === 'processing',
-    nextPage,
-    prevPage,
-    emptyTableText: getEmptyTableText(fetchAttempt.status),
-    customSort: {
-      dir: agentFilter.sort?.dir,
-      fieldName: agentFilter.sort?.fieldName,
-      onSort: updateSort,
-    },
+    kubes: fetchAttempt.data?.agentsList.map(makeKube) || [],
+    ...serversideResources,
   };
 }
